+++ tests/testthat/in/style-files-ignore-fail.R
@@ -1,0 +1,7 @@
+# styler: off
+1+1
+# styler: on
+
+12 + 39
+
+# 1 styler: on